dac_sine I2C, PLLI2S, TLV register bug fixes
- Fixed a problem which prevented reading from I2C devices.
- Set the PLLI2S properly, as listed in CMSIS documentation.
- Fixed some TLV registers responsible for muting the chip, and added
additional TLV register writes to unmute the chip.

+++ TLV/src/ringbuffer.c
@@ -43,7 +43,7 @@ int32_t get(RingBuffer* buffer)
 	//check space
 	if (hasElement(buffer) != 0)
 	{	//get character from the  buffer
-		uint16_t element = buffer->buffer[buffer->tail];
+		int32_t element = buffer->buffer[buffer->tail];
 		//adjusting get
 		buffer->tail = (buffer->tail+1) % BUF_SIZE;
 		return element;
